Rename to stationService.js
